refactor(author): move add, update, delete logic to admin controller/services
