+++ backend/src/services/core/author_services.js
@@ -37,35 +37,3 @@ export async function getAuthorService(id) {
   }
   return author;
 }
-
-export async function addAuthorService(author) {
-  const { penName, bio } = author;
-  const newAuthor = new Author({
-    penName,
-    bio,
-  });
-  await newAuthor.save();
-  return newAuthor;
-}
-
-export async function updateAuthorService(author) {
-  const { authorCode, penName, bio } = author;
-  const updateAuthor = await Author.findOneAndUpdate(
-    { authorCode: authorCode },
-    {
-      penName,
-      bio,
-    },
-    { new: true },
-  );
-  if (!updateAuthor) {
-    const error = new Error("Author not found");
-    error.status = 404;
-    throw error;
-  }
-  return updateAuthor;
-}
-
-export async function deleteAuthorService(author) {
-  return await Author.findByIdAndDelete(author);
-}
